pyproject.toml
[project]
name = "cvproject"
version = "0.1.0"
description = "DTEAM - Django Developer Practical Test"
authors = [
    {name = "Shams",email = "shamsulloismatov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==4.2.10)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-nested-routers (>=0.94.1,<0.95.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-weasyprint (>=2.4.0,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-celery-results (>=2.5.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"

[tool.poetry.group.production.dependencies]
gunicorn = "^23.0.0"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
skip-string-normalization = false
skip-magic-trailing-comma = true